Devices/Storage/DevLsiLogicSCSI: Use the default devie critical section to improve behavior with SMP guests and some other smaller fixes, bugref:9767 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@85102 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Storage/DevLsiLogicSCSI.cpp
@@ -1983,7 +1983,7 @@ static size_t lsilogicSgBufWalker(PPDMDEVINS pDevIns, PLSILOGICREQ pLsiReq,
                 && SGEntry.Simple32.fEndOfBuffer)
                 return cbCopied - RT_MIN(cbSkip, cbCopied);
 
-            uint32_t cbCopyThis           = RT_MIN(SGEntry.Simple32.u24Length, cbCopy);
+            size_t   cbCopyThis           = RT_MIN(SGEntry.Simple32.u24Length, cbCopy);
             RTGCPHYS GCPhysAddrDataBuffer = SGEntry.Simple32.u32DataBufferAddressLow;
 
             if (SGEntry.Simple32.f64BitAddress)
